feat: update product card and shop screen
- Updated the product card widget to display product information including image, rating, review count, brand, title, and price.
- Integrated product data from a provider into the shop screen to populate the product grid.
- Implemented loading and error handling in the shop screen.
- Added a search bar to filter products by name.
- Modified data models (Variant, Product) to support double type for prices.
- Updated the product API to fetch and parse product data, including variants and ratings.

+++ lib/screens/home/screen/dashboard_screen.dart
@@ -132,8 +132,8 @@ class DashboardScreen extends StatelessWidget {
               discount: ['-5%', '-5%', '-10%'][index],
               brand: 'Apple',
               title: ['Iphone 15 Pro', 'Iphone 14 Pro', 'Iphone 15 Pro Max'][index],
-              originalPrice: [400, 500, 600][index],
-              discountedPrice: [380, 450, 540][index],
+              originalPrice: [400, 500, 600][index].toDouble(),
+              discountedPrice: [380, 450, 540][index].toDouble(),
             );
           },
           separatorBuilder: (context, index) {
